+++ workspaces/init.go
@@ -2,8 +2,10 @@ package workspaces
 
 import (
 	"context"
+	"fmt"
 	"os"
 	"os/exec"
+	"strings"
 )
 
 func Init(ctx context.Context, toolName string) error {
@@ -14,13 +16,14 @@ func Init(ctx context.Context, toolName string) error {
 	case "terraform":
 		process = "terraform"
 	case "opentofu":
-		process = "opentofu"
+		process = "tofu"
 	}
 
 	args := []string{
 		"init",
 		"-reconfigure",
 	}
+	fmt.Printf("Running `%s %s`\n", process, strings.Join(args, " "))
 	cmd := exec.CommandContext(ctx, process, args...)
 	cmd.Stderr = os.Stderr
 	cmd.Stdout = os.Stdout